# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/posts": {
      "get": {
        "operationId": "PostsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "PostsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{slug}": {
      "get": {
        "operationId": "PostsController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "description": "pt (default), en, or es — resolves that locale's own slug column first, falling back to the default \"slug\" when the post has no translated slug of its own.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}": {
      "patch": {
        "operationId": "PostsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "operationId": "PostsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "tags": [
          "posts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vertex API",
    "description": "API documentation for the Vertex backend platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
...
  }
}